+++ src/typemap/src/datatype_conversion.rs
@@ -9,46 +9,10 @@
 use cage::{get_cage, memory::memory::translate_vmmap_addr};
 use fdtables;
 use std::error::Error;
-use std::str::Utf8Error;
 use sysdefs::constants::err_const::{syscall_error, Errno};
 use sysdefs::constants::lind_platform_const::{MAX_CAGEID, PATH_MAX};
 use sysdefs::data::fs_struct::{SigactionStruct, SigsetType, ITimerVal};
 use sysdefs::constants::lind_platform_const::{UNUSED_ARG, UNUSED_ID, UNUSED_NAME};
-
-/// This function provides two operations: first, it translates path pointer address from WASM environment
-/// to kernel system address; then, it adjusts the path from user's perspective to host's perspective,
-/// which is adding `LIND_ROOT` before the path arguments. Considering actual syscall implementation
-/// logic needs to pass string pointer to underlying rust libc, so this function will return `CString`
-/// always using arg_cageid to translate. (TODO: the logic here might be different according to 3i/grate
-/// implementation)
-///     - If arg_cageid != cageid: this call is sent by grate. We need to translate according to cage
-///     - If arg_cageid == cageid: this call is sent by cage, we can use either one
-///
-/// ## Arguments:
-///     - cageid: required to do address translation for path pointer
-///     - path_arg: the path pointer with address and contents from user's perspective. Address is
-///                 32-bit (because of WASM feature).
-///
-/// ## Returns:
-///     - c_path: a `CString` variable stores the path from host's perspective
-///     - will return error if total length exceed the MAX_PATH (which is 4096). We use `Box<dyn Error>` here to
-///      let upper functions do error handling. (ie: we want to )
-pub unsafe fn charstar_to_ruststr<'a>(cstr: *const i8) -> Result<&'a str, Utf8Error> {
-    std::ffi::CStr::from_ptr(cstr as *const _).to_str() //returns a result to be unwrapped later
-}
-
-pub fn get_cstr<'a>(arg: u64) -> Result<&'a str, i32> {
-    let ptr = arg as *const i8;
-    if !ptr.is_null() {
-        if let Ok(data) = unsafe { charstar_to_ruststr(ptr) } {
-            return Ok(data);
-        } else {
-            return Err(-1);
-        }
-    }
-
-    return Err(-1);
-}
 
 /// `sc_unusedarg()` is the security check function used to validate all unused args. This
 /// will return true in default mode, and check if `arg` with `arg_cageid` are all null in
@@ -359,7 +323,7 @@ pub fn sc_convert_addr_to_host(addr_arg: u64, addr_arg_cageid: u64, cageid: u64)
 /// # Returns
 /// * `Some(&SigactionStruct)` if the pointer is nonzero and translation succeeds.
 /// * `None` if `act_arg == 0`.
-pub fn sc_convert_SigactionStruct<'a>(
+pub fn sc_convert_sigactionStruct<'a>(
     act_arg: u64,
     act_arg_cageid: u64,
     cageid: u64, 
@@ -402,7 +366,7 @@ pub fn sc_convert_SigactionStruct<'a>(
 /// # Returns
 /// * `Some(&mut SigactionStruct)` if the pointer is nonzero and translation succeeds.
 /// * `None` if `act_arg == 0`.
-pub fn sc_convert_SigactionStruct_mut<'a>(
+pub fn sc_convert_sigactionStruct_mut<'a>(
     act_arg: u64,
     act_arg_cageid: u64,
     cageid: u64,
